% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mq_adjustment}
\alias{mq_adjustment}
\title{Marginal Quantile-Adjustment for Fair and Interpretable Policy Learning}
\usage{
mq_adjustment(
  vars,
  sens,
  seed = 123456,
  ties.method = "random",
  quantile.type = 4
)
}
\arguments{
\item{vars}{Numeric matrix or data.frame of variables to be adjusted (observations in rows, variables in columns).}

\item{sens}{Matrix or data.frame of sensitive attributes. Must have the same number of rows as `vars`.}

\item{seed}{Integer scalar for reproducible random tie-breaking.}

\item{ties.method}{Character string for ranking ties. One of "random", "average", "first", "last", "max", "min".}

\item{quantile.type}{An integer from \code{1} to \code{9}, or the character string \code{"reshuffled"} selecting the quantile
algorithms. See \code{\link[stats]{quantile}} for details. \code{"reshuffled"} provides an alternative algorithm that keeps
the exact moments of the original distribution.}
}
\value{
A list with two data.frames:
  \describe{
    \item{vars_cdf}{CDF-adjusted variables denoted with suffix `_cdf`.}
    \item{vars_mq}{MQ-adjusted variables denoted with suffix `_mq`.}
  }
}
\description{
Computes fairness-adjusted variables by MQ-adjustment
}
