% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot.prob_split_tree_list}
\alias{plot.prob_split_tree_list}
\title{Plot a List of Probabilistic Split Trees.}
\usage{
\method{plot}{prob_split_tree_list}(tree_list, sens_names = NULL, leaf.labels = NULL)
}
\arguments{
\item{sens_names}{The variables names of the sensitive attributes.}

\item{leaf.labels}{An optional character vector of leaf labels for each treatment.}

\item{tree}{A fitted object of class `'prob_split_tree_list'`.}
}
\description{
Plot a List of Probabilistic Split Trees.
}
