% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{predict.prob_split_tree}
\alias{predict.prob_split_tree}
\title{Predict from a Probabilistic Split Tree}
\usage{
\method{predict}{prob_split_tree}(
  tree,
  vars,
  tree_cdf = NULL,
  vars_cdf = NULL,
  type = "action.id",
  seed = 123456
)
}
\arguments{
\item{tree}{A fitted object of class `'prob_split_tree'`.}

\item{vars}{A data.frame or matrix of the decision variables.}

\item{tree_cdf}{Optional. A corresponding `policy_tree` object fitted on CDF-adjusted decision variables.}

\item{vars_cdf}{Optional. A data.frame or matrix of CDF-adjusted decision variables (same structure as `vars`).}

\item{type}{Character. One of `"action.id"` (default) or `"leaf.id"`. Determines the type of prediction returned.}

\item{seed}{Integer. Random seed used to resolve probabilistic splits when exact info is missing. Default is 123456.}
}
\value{
A numeric vector of predicted action or leaf IDs for each observation.
}
\description{
Predict from a Probabilistic Split Tree
}
