% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print.prob_split_tree}
\alias{print.prob_split_tree}
\title{Print a Probabilistic Split Tree.}
\usage{
\method{print}{prob_split_tree}(x, ...)
}
\arguments{
\item{x}{A fitted object of class `'prob_split_tree'`.}

\item{...}{Currently ignored.}
}
\description{
Print a Probabilistic Split Tree.
}
