% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print.prob_split_tree_list}
\alias{print.prob_split_tree_list}
\title{Print a List of Probabilistic Split Trees.}
\usage{
\method{print}{prob_split_tree_list}(tree_list, sens_names = NULL)
}
\arguments{
\item{sens_names}{The variables names of the sensitive attributes.}

\item{tree}{A fitted object of class `'prob_split_tree_list'`.}
}
\description{
Print a List of Probabilistic Split Trees.
}
