% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot.prob_split_tree}
\alias{plot.prob_split_tree}
\title{Plot a Probabilistic Split Tree.}
\usage{
\method{plot}{prob_split_tree}(x, ...)
}
\arguments{
\item{x}{A fitted object of class `'prob_split_tree'`.}

\item{...}{Additional arguments passed to the plotting function.
Currently supports:
\describe{
  \item{`leaf.labels`}{An optional character vector of leaf labels for each treatment.}
}}
}
\description{
Plot a Probabilistic Split Tree.
}
