% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot.prob_split_tree_list}
\alias{plot.prob_split_tree_list}
\title{Plot a List of Probabilistic Split Trees.}
\usage{
\method{plot}{prob_split_tree_list}(x, ...)
}
\arguments{
\item{x}{A fitted object of class `'prob_split_tree_list'`.}

\item{...}{Additional arguments passed to the plotting function.
Currently supports:
\describe{
  \item{`sens_names`}{An optional character vector of variables names of the sensitive attributes.}
  \item{`leaf.labels`}{An optional character vector of leaf labels for each treatment.}
}}
}
\description{
Plot a List of Probabilistic Split Trees.
}
