% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{print.prob_split_tree_list}
\alias{print.prob_split_tree_list}
\title{Print a List of Probabilistic Split Trees.}
\usage{
\method{print}{prob_split_tree_list}(x, ...)
}
\arguments{
\item{x}{A fitted object of class `'prob_split_tree_list'`.}

\item{...}{Additional arguments passed to the printing function.
Currently supports:
\describe{
  \item{`sens_names`}{Character vector, the variables names of the sensitive attributes.}
}}
}
\description{
Print a List of Probabilistic Split Trees.
}
